fix: all table models in the same file
